+++ CHANGELOG.md
@@ -10,6 +10,8 @@
     - Bugfixes
       - fixed a bug where polyline decoding on a defective polyline could end up in out-of-bound access on a vector
       - fixed compile errors in tile unit-test framework
+    - Debug Tiles
+      - Added support for turn penalties
 
 # 5.4.0
   - Changes from 5.3.0